pyproject.toml
[tool.poetry]
name = "armallm"
version = "0.1.0"
description = ""
authors = ["Ivan"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
tqdm = "^4.67.1"
beautifulsoup4 = "^4.12.3"
pandas = "^2.2.3"
playwright = "^1.49.1"
pyyaml = "^6.0.2"
rich = "^13.9.4"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
